teacher-attendance-app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.1
Werkzeug==2.3.6